Don't create database automatically for entity framework based plugins
Removed call for context.Database.EnsureCreatedAsync(); so that database is not created automatically. A new configuration flag 'AutoMigrate' is added with default value of false that can apply migrataions automatically. This will be useful for dev/ci-cd environments and to quickly spin up containers. Sql migration scripts should be used for actual production usecase.